payment summary filter button color and delete button popup

+++ src/pages/Payments/PaymentSummary.tsx
@@ -9,6 +9,7 @@ import { MdOutlineFilterAltOff } from "react-icons/md";
 import { RiDeleteBin7Line } from "react-icons/ri";
 import { useLocation, useNavigate } from "react-router-dom";
 import PaymentTable from "./PaymentTable";
+import Swal from "sweetalert2";
 const { Search } = Input;
 
 const PaymentSummary: React.FC = () => {
@@ -21,7 +22,7 @@ const PaymentSummary: React.FC = () => {
     page: 1,
     limit: 10,
   });
-  const [deleteMany] = useDeleteManyPaymentMutation();
+  const [deleteMany, { isSuccess }] = useDeleteManyPaymentMutation();
 
   // states
   const [search, setSearch] = useState(queryParams.search as string);
@@ -38,9 +39,34 @@ const PaymentSummary: React.FC = () => {
     navigate(location.pathname, { replace: true });
   };
 
-  const handleDeleteMany = () => {
-    deleteMany({ ids: selectedRowKeys });
+  const handleDeleteMany = async () => {
+    const result = await Swal.fire({
+      // title: "",
+      text: "Are you sure want to remove this?",
+      showCancelButton: true,
+      confirmButtonText: "Confirm",
+    });
+
+    if (result.isConfirmed) {
+      try {
+        deleteMany({ ids: selectedRowKeys });
+      } catch (error) {
+        console.error(
+          error instanceof Error ? error.message : "An unknown error occurred."
+        );
+      }
+    }
   };
+
+  if (isSuccess) {
+    Swal.fire({
+      title: "",
+      text: "Payment Deleted Successfully.",
+      icon: "success",
+      timer: 2000,
+      showConfirmButton: false,
+    });
+  }
 
   return (
     <>
@@ -57,11 +83,10 @@ const PaymentSummary: React.FC = () => {
         />
 
         <Button
-          variant="solid"
-          color="purple"
+          type="primary"
+          className="!bg-gray-300 !text-black hover:!bg-gray-400"
           title="Clear All Filters"
           onClick={handleClearQuery}
-          danger
           icon={<MdOutlineFilterAltOff />}
         />
         <Button
